requirements.txt
discord.py==1.6.0
psycopg2==2.8.5
pytest==8.2.0
python-dotenv==1.0.1